+++ component/src/test/java/org/ballerinalang/kafka/test/consumer/KafkaConsumerManualCommitTest.java
@@ -42,8 +42,8 @@ import java.io.IOException;
 import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
 
-import static org.ballerinalang.kafka.test.utils.Utils.KAFKA_BROKER_PORT;
-import static org.ballerinalang.kafka.test.utils.Utils.ZOOKEEPER_PORT_1;
+import static org.ballerinalang.kafka.test.utils.Constants.KAFKA_BROKER_PORT;
+import static org.ballerinalang.kafka.test.utils.Constants.ZOOKEEPER_PORT_1;
 import static org.ballerinalang.kafka.util.KafkaConstants.KAFKA_NATIVE_PACKAGE;
 
 /**
