+++ component/src/test/java/org/ballerinalang/kafka/test/consumer/KafkaConsumerSubscribePartitionRebalanceTest.java
@@ -37,8 +37,8 @@ import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 import static org.awaitility.Awaitility.await;
-import static org.ballerinalang.kafka.test.utils.Utils.KAFKA_BROKER_PORT;
-import static org.ballerinalang.kafka.test.utils.Utils.ZOOKEEPER_PORT_1;
+import static org.ballerinalang.kafka.test.utils.Constants.KAFKA_BROKER_PORT;
+import static org.ballerinalang.kafka.test.utils.Constants.ZOOKEEPER_PORT_1;
 
 @Test(singleThreaded = true)
 public class KafkaConsumerSubscribePartitionRebalanceTest {
