+++ component/src/test/java/org/ballerinalang/kafka/test/producer/KafkaProducerTest.java
@@ -38,8 +38,8 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 
-import static org.ballerinalang.kafka.test.utils.Utils.KAFKA_BROKER_PORT;
-import static org.ballerinalang.kafka.test.utils.Utils.ZOOKEEPER_PORT_1;
+import static org.ballerinalang.kafka.test.utils.Constants.KAFKA_BROKER_PORT;
+import static org.ballerinalang.kafka.test.utils.Constants.ZOOKEEPER_PORT_1;
 
 /**
  * Test cases for ballerina.net.kafka producer connector.
